Split workflow contract YAML decoder into focused files by concern
